+++ CHANGELOG.md
@@ -7,9 +7,9 @@
  * Gradle 5 compatible (up to 5.6.4) (#84)
  * Use Gradle 5.6.4 to build (#84)
  * Use Kotlin DSL 5.6.4 (#84)
- * Android Gradle Plugin 3.5.3 compatible (#85)
+ * Android Gradle Plugin 3.5.3 compatible (#86)
 
-## Fixes
+### Fixes
  * Make sure nothing is logged when lint task is disabled.